% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.r
\name{caption}
\alias{caption}
\title{Caption}
\usage{
caption(template, glyphs, out = NULL)
}
\arguments{
\item{template}{text with placeholders for each glyph}

\item{glyphs}{list of glyphs with mapping metadata}

\item{out}{optional output file name}
}
\value{
markdown
}
\description{
Caption
}

