% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captionGrob.r
\name{captionGrob}
\alias{captionGrob}
\title{Caption rendering with grid graphics}
\usage{
captionGrob(..., width = unit(4, "in"))
}
\arguments{
\item{...}{grobs}

\item{width}{width of the caption as absolute grid unit}
}
\value{
grob
}
\description{
Caption rendering with grid graphics
}

