% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{extract_legend_grobs}
\alias{extract_legend_grobs}
\title{Extract legend grobs}
\usage{
extract_legend_grobs(p)
}
\arguments{
\item{p}{ggplot}
}
\value{
list with mapping metadata
}
\description{
Extract legend grobs
}

